requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

bs4==0.0.2
mysql-connector==2.2.9
openai==0.28.0
python-dotenv==1.0.1
requests==2.32.3
